+++ internal/billing/domain/events/billing_requested.go
@@ -3,10 +3,12 @@ package events
 import "time"
 
 type BillingRequested struct {
-	BillingID   string
-	InvoiceID   string
-	AmountCents int64
-	DueDate     time.Time
+	BillingID     string
+	InvoiceID     string
+	CustomerID    string
+	AmountCents   int64
+	DueDate       time.Time
+	AttemptNumber int
 }
 
 func (BillingRequested) Name() string {